requirements.txt
numpy
pandas
pyarrow
scikit-learn
tqdm
pyyaml
huggingface_hub
sentence-transformers
openai
matplotlib
